Fix #359 Move package wollmux.ui to UNOHelper
Was done with https://github.com/WollMux/UNOHelper/commit/bf490c22882352b02d93db5d67afc99c2c2460b4
