+++ core/src/main/java/de/muenchen/allg/itd51/wollmux/form/config/FormConfig.java
@@ -33,8 +33,8 @@ import org.slf4j.LoggerFactory;
 import de.muenchen.allg.itd51.wollmux.config.ConfigThingy;
 import de.muenchen.allg.itd51.wollmux.config.NodeNotFoundException;
 import de.muenchen.allg.itd51.wollmux.form.model.FormModelException;
-import de.muenchen.allg.itd51.wollmux.ui.UIElementConfig;
 import de.muenchen.allg.itd51.wollmux.util.L;
+import de.muenchen.allg.itd51.wollmux.ui.UIElementConfig;
 
 /**
  * Description of a form.
