+++ core/src/main/java/de/muenchen/allg/itd51/wollmux/form/sidebar/FormSidebarPanel.java
@@ -68,16 +68,16 @@ import de.muenchen.allg.itd51.wollmux.form.config.FormConfig;
 import de.muenchen.allg.itd51.wollmux.form.config.TabConfig;
 import de.muenchen.allg.itd51.wollmux.form.model.Control;
 import de.muenchen.allg.itd51.wollmux.form.model.FormModel;
-import de.muenchen.allg.itd51.wollmux.ui.GuiFactory;
-import de.muenchen.allg.itd51.wollmux.ui.HTMLElement;
+import de.muenchen.allg.ui.layout.ControlLayout;
+import de.muenchen.allg.ui.layout.FixedVerticalLayout;
+import de.muenchen.allg.ui.layout.HorizontalLayout;
+import de.muenchen.allg.ui.layout.Layout;
+import de.muenchen.allg.ui.layout.TabLayout;
+import de.muenchen.allg.ui.layout.VerticalLayout;
+import de.muenchen.allg.ui.GuiFactory;
+import de.muenchen.allg.ui.HTMLElement;
 import de.muenchen.allg.itd51.wollmux.ui.UIElementConfig;
 import de.muenchen.allg.itd51.wollmux.ui.UIElementType;
-import de.muenchen.allg.itd51.wollmux.ui.layout.ControlLayout;
-import de.muenchen.allg.itd51.wollmux.ui.layout.FixedVerticalLayout;
-import de.muenchen.allg.itd51.wollmux.ui.layout.HorizontalLayout;
-import de.muenchen.allg.itd51.wollmux.ui.layout.Layout;
-import de.muenchen.allg.itd51.wollmux.ui.layout.TabLayout;
-import de.muenchen.allg.itd51.wollmux.ui.layout.VerticalLayout;
 import de.muenchen.allg.util.UnoConfiguration;
 import de.muenchen.allg.util.UnoProperty;
