+++ core/src/test/java/de/muenchen/allg/itd51/wollmux/ui/layout/DummyLayout.java
@@ -28,6 +28,8 @@ import org.apache.commons.lang3.tuple.Pair;
 
 import com.sun.star.awt.Rectangle;
 
+import de.muenchen.allg.ui.layout.Layout;
+
 public class DummyLayout implements Layout
 {
   private boolean visible;
