+++ pom.xml
@@ -237,7 +237,7 @@
       <dependency>
         <groupId>de.muenchen</groupId>
         <artifactId>unohelper</artifactId>
-        <version>4.0.3-SNAPSHOT</version>
+        <version>4.1.0-SNAPSHOT</version>
       </dependency>
       <dependency>
         <groupId>org.openjfx</groupId>
